requirements.txt
requests
psycopg2-binary
python-dotenv
openai
pandas
numpy
beautifulsoup4
lxml
playwright
fastapi
uvicorn

# Dependencias para MongoDB y Embeddings
pymongo>=4.6.0
pydantic>=2.0.0
openai>=1.0.0
numpy>=1.24.0
